+++ tests/Tests/Parametizer/Parsing/ParsingTest.php
@@ -239,8 +239,12 @@ class ParsingTest extends TestCaseAbstract {
      * @covers Parametizer::run()
      * @covers Parser::read()
      */
-    public function testParsingManyFlagsSuccess(string $script, string $parametersString, array $expectedValues): void {
-        $result = static::assertNoErrorsOutput($script, $parametersString);
+    public function testParsingManyFlagsSuccess(
+        string $scriptPath,
+        string $parametersString,
+        array $expectedValues,
+    ): void {
+        $result = static::assertNoErrorsOutput($scriptPath, $parametersString);
 
         $resultJsonString = $result->getStdOut();
         $actualValues     = json_decode($resultJsonString, true);
@@ -254,29 +258,29 @@ class ParsingTest extends TestCaseAbstract {
         return [
             /** Three short name flags at once {@see Parser::read()} */
             'many-short-flags-1' => [
-                'script'           => __DIR__ . '/scripts/many-flags.php',
+                'scriptPath'       => __DIR__ . '/scripts/many-flags.php',
                 'parametersString' => '-xyz',
                 'expectedValues'   => ['flag-x' => true, 'flag-y' => true, 'flag-z' => true],
             ],
             'many-short-flags-2' => [
-                'script'           => __DIR__ . '/scripts/many-flags.php',
+                'scriptPath'       => __DIR__ . '/scripts/many-flags.php',
                 'parametersString' => '-xy -z',
                 'expectedValues'   => ['flag-x' => true, 'flag-y' => true, 'flag-z' => true],
             ],
 
             /** Flags and options with short names {@see Parser::read()} */
             'option-with-short-name-and-value-1' => [
-                'script'           => __DIR__ . '/scripts/many-flags.php',
+                'scriptPath'       => __DIR__ . '/scripts/many-flags.php',
                 'parametersString' => '-xy -oz',
                 'expectedValues'   => ['flag-x' => true, 'flag-y' => true, 'option' => 'z'],
             ],
             'option-with-short-name-and-value-2' => [
-                'script'           => __DIR__ . '/scripts/many-flags.php',
+                'scriptPath'       => __DIR__ . '/scripts/many-flags.php',
                 'parametersString' => '-xyoz',
                 'expectedValues'   => ['flag-x' => true, 'flag-y' => true, 'option' => 'z'],
             ],
             'option-with-short-name-and-value-3' => [
-                'script'           => __DIR__ . '/scripts/many-flags.php',
+                'scriptPath'       => __DIR__ . '/scripts/many-flags.php',
                 'parametersString' => '-xozy',
                 'expectedValues'   => ['flag-x' => true, 'option' => 'zy'],
             ],
@@ -291,7 +295,6 @@ class ParsingTest extends TestCaseAbstract {
      *  1) a script help parts are printed for all missing required parameters and parameters with invalid values;
      *  2) the `--help` parameter is always printed (so a user could know how to see a full help page).
      *
-     * @param string[] $helpSubstrings
      * @covers HelpGenerator::getUsageForParseErrorException()
      * @covers Parametizer::run()
      * @covers CliRequestProcessor::registerArgument()
@@ -300,12 +303,11 @@ class ParsingTest extends TestCaseAbstract {
      * @covers CliRequestProcessor::validate()
      */
     public function testParseErrorsWithHelp(
-        string $script,
+        string $scriptPath,
         string $parametersString,
-        string $errorOutput,
-        array $helpSubstrings,
+        string $expectedErrorOutput,
     ): void {
-        static::assertParseErrorOutputWithHelp($script, $errorOutput, $parametersString, $helpSubstrings);
+        static::assertFullErrorOutput($scriptPath, $expectedErrorOutput, $parametersString);
     }
 
     /**
@@ -314,131 +316,254 @@ class ParsingTest extends TestCaseAbstract {
     public static function provideParseErrorsWithHelp(): array {
         return [
             'required-params-missed' => [
-                'script'           => __DIR__ . '/scripts/lots-of-params.php',
-                'parametersString' => '',
-                'errorOutput'      => 'Need more parameters',
-                'helpSubstrings'   => [
-                    '--opt-required=…   Required option',
-                    '<arg-required>     Required argument',
-                ],
+                'scriptPath'          => __DIR__ . '/scripts/lots-of-params.php',
+                'parametersString'    => '',
+                'expectedErrorOutput' => <<<STDERR_OUTPUT
+Need more parameters
+
+
+  --help             Show full help page.
+
+  --opt-required=…   Required option: pick one from the list
+  (required)         Allowed values:
+                      - black A pile of books
+                      - pink  A heap of ponies
+                      - white
+                      - 5     Give me "five"!
+
+  <arg-required>     Required argument
+  (required)
+
+STDERR_OUTPUT,
             ],
             'required-option-missed' => [
-                'script'           => __DIR__ . '/scripts/lots-of-params.php',
-                'parametersString' => 'arg',
-                'errorOutput'      => 'Need a value for --opt-required',
-                'helpSubstrings'   => [
-                    '--opt-required=…   Required option',
-                ],
+                'scriptPath'          => __DIR__ . '/scripts/lots-of-params.php',
+                'parametersString'    => 'arg',
+                'expectedErrorOutput' => <<<STDERR_OUTPUT
+Need a value for --opt-required
+
+
+  --help             Show full help page.
+
+  --opt-required=…   Required option: pick one from the list
+  (required)         Allowed values:
+                      - black A pile of books
+                      - pink  A heap of ponies
+                      - white
+                      - 5     Give me "five"!
+
+STDERR_OUTPUT,
             ],
             'required-argument-missed' => [
-                'script'           => __DIR__ . '/scripts/lots-of-params.php',
-                'parametersString' => '--opt-required=5',
-                'errorOutput'      => 'Need more parameters',
-                'helpSubstrings'   => [
-                    '<arg-required>   Required argument',
-                ],
+                'scriptPath'          => __DIR__ . '/scripts/lots-of-params.php',
+                'parametersString'    => '--opt-required=5',
+                'expectedErrorOutput' => <<<STDERR_OUTPUT
+Need more parameters
+
+
+  --help           Show full help page.
+
+  <arg-required>   Required argument
+  (required)
+
+STDERR_OUTPUT,
             ],
             'required-several-options-missed' => [
-                'script'           => __DIR__ . '/scripts/several-required-options.php',
-                'parametersString' => '',
-                'errorOutput'      => 'Need values for --option1 (-f), --option2, --option3 (-t)',
-                'helpSubstrings'   => [
-                    '-f …, --option1=…   First option',
-                    '--option2=…         Second option',
-                    '-t …, --option3=…   Third option',
-                ],
+                'scriptPath'          => __DIR__ . '/scripts/several-required-options.php',
+                'parametersString'    => '',
+                'expectedErrorOutput' => <<<STDERR_OUTPUT
+Need values for --option1 (-f), --option2, --option3 (-t)
+
+
+  --help              Show full help page.
+
+  -f …, --option1=…   First option
+  (required)
+
+  --option2=…         Second option
+  (required)
+
+  -t …, --option3=…   Third option
+  (required)
+
+STDERR_OUTPUT,
             ],
 
             'argument-invalid-value' => [
-                'script'           => __DIR__ . '/scripts/lots-of-params.php',
-                'parametersString' => '--opt-required=5 arg D',
-                'errorOutput'      => "Incorrect value 'D' for argument <arg-optional>",
-                'helpSubstrings'   => [
-                    '<arg-optional>   Optional argument: pick one from the list',
-                ],
+                'scriptPath'          => __DIR__ . '/scripts/lots-of-params.php',
+                'parametersString'    => '--opt-required=5 arg D',
+                'expectedErrorOutput' => <<<STDERR_OUTPUT
+Incorrect value 'D' for argument <arg-optional>
+
+
+  --help           Show full help page.
+
+  <arg-optional>   Optional argument: pick one from the list
+                   Allowed values: A, B, C
+                   Default: B
+
+STDERR_OUTPUT,
             ],
 
             'option-without-value-1' => [
-                'script'           => __DIR__ . '/scripts/lots-of-params.php',
-                'parametersString' => '--opt-required arg',
-                'errorOutput'      => 'No value for option --opt-required',
-                'helpSubstrings'   => [
-                    '--opt-required=…   Required option',
-                ],
+                'scriptPath'          => __DIR__ . '/scripts/lots-of-params.php',
+                'parametersString'    => '--opt-required arg',
+                'expectedErrorOutput' => <<<STDERR_OUTPUT
+No value for option --opt-required
+
+
+  --help             Show full help page.
+
+  --opt-required=…   Required option: pick one from the list
+  (required)         Allowed values:
+                      - black A pile of books
+                      - pink  A heap of ponies
+                      - white
+                      - 5     Give me "five"!
+
+STDERR_OUTPUT,
             ],
             'option-without-value-2' => [
-                'script'           => __DIR__ . '/scripts/lots-of-params.php',
-                'parametersString' => '--opt-required= arg',
-                'errorOutput'      => 'No value for option --opt-required',
-                'helpSubstrings'   => [
-                    '--opt-required=…   Required option',
-                ],
+                'scriptPath'          => __DIR__ . '/scripts/lots-of-params.php',
+                'parametersString'    => '--opt-required= arg',
+                'expectedErrorOutput' => <<<STDERR_OUTPUT
+No value for option --opt-required
+
+
+  --help             Show full help page.
+
+  --opt-required=…   Required option: pick one from the list
+  (required)         Allowed values:
+                      - black A pile of books
+                      - pink  A heap of ponies
+                      - white
+                      - 5     Give me "five"!
+
+STDERR_OUTPUT,
             ],
 
             'option-duplicate-value' => [
-                'script'           => __DIR__ . '/scripts/lots-of-params.php',
-                'parametersString' => '--opt-required=white arg --opt-required=pink',
-                'errorOutput'      => "Duplicate option --opt-required (with value 'pink'); already registered value: 'white'",
-                'helpSubstrings'   => [
-                    '--opt-required=…   Required option',
-                ],
+                'scriptPath'          => __DIR__ . '/scripts/lots-of-params.php',
+                'parametersString'    => '--opt-required=white arg --opt-required=pink',
+                'expectedErrorOutput' => <<<STDERR_OUTPUT
+Duplicate option --opt-required (with value 'pink'); already registered value: 'white'
+
+
+  --help             Show full help page.
+
+  --opt-required=…   Required option: pick one from the list
+  (required)         Allowed values:
+                      - black A pile of books
+                      - pink  A heap of ponies
+                      - white
+                      - 5     Give me "five"!
+
+STDERR_OUTPUT,
             ],
             'array-option-duplicate-value' => [
-                'script'           => __DIR__ . '/scripts/lots-of-params.php',
-                'parametersString' => '--opt-required=5 arg -l100 -l800 -l100',
-                'errorOutput'      => "Duplicate value '100' for option --opt-list (-l); already registered values: '100', '800'",
-                'helpSubstrings'   => [
-                    '-l …, --opt-list=…   List of values',
-                ],
+                'scriptPath'          => __DIR__ . '/scripts/lots-of-params.php',
+                'parametersString'    => '--opt-required=5 arg -l100 -l800 -l100',
+                'expectedErrorOutput' => <<<STDERR_OUTPUT
+Duplicate value '100' for option --opt-list (-l); already registered values: '100', '800'
+
+
+  --help               Show full help page.
+
+  -l …, --opt-list=…   List of values
+                       Allowed values: 100, 150, 200, 250, 300, 350, 400, 450, 500, 550, 600, 650, 700, 750, 800
+                       (multiple values allowed)
+
+STDERR_OUTPUT,
             ],
 
             'option-invalid-value' => [
-                'script'           => __DIR__ . '/scripts/lots-of-params.php',
-                'parametersString' => '--opt-required=blue arg',
-                'errorOutput'      => "Incorrect value 'blue' for option --opt-required",
-                'helpSubstrings'   => [
-                    '--opt-required=…   Required option',
-                ],
+                'scriptPath'          => __DIR__ . '/scripts/lots-of-params.php',
+                'parametersString'    => '--opt-required=blue arg',
+                'expectedErrorOutput' => <<<STDERR_OUTPUT
+Incorrect value 'blue' for option --opt-required
+
+
+  --help             Show full help page.
+
+  --opt-required=…   Required option: pick one from the list
+  (required)         Allowed values:
+                      - black A pile of books
+                      - pink  A heap of ponies
+                      - white
+                      - 5     Give me "five"!
+
+STDERR_OUTPUT,
             ],
 
             'flag-with-value' => [
-                'script'           => __DIR__ . '/scripts/lots-of-params.php',
-                'parametersString' => '--opt-required=5 arg --flag1=test',
-                'errorOutput'      => 'The flag --flag1 (-f) can not have a value',
-                'helpSubstrings'   => [
-                    '-f, --flag1   Some flag',
-                ],
+                'scriptPath'          => __DIR__ . '/scripts/lots-of-params.php',
+                'parametersString'    => '--opt-required=5 arg --flag1=test',
+                'expectedErrorOutput' => <<<STDERR_OUTPUT
+The flag --flag1 (-f) can not have a value
+
+
+  --help        Show full help page.
+
+  -f, --flag1   Some flag
+
+STDERR_OUTPUT,
             ],
 
             'non-configured-extra-argument' => [
-                'script'           => __DIR__ . '/scripts/two-arguments.php',
-                'parametersString' => 'asd fgh unknown1 unknown2',
-                'errorOutput'      => "Too many arguments, starting with 'unknown1'",
-                'helpSubstrings'   => [
-                    '<argument-one>   First argument',
-                    '<argument-two>   Second argument'
-                ],
+                'scriptPath'          => __DIR__ . '/scripts/two-arguments.php',
+                'parametersString'    => 'asd fgh unknown1 unknown2',
+                'expectedErrorOutput' => <<<STDERR_OUTPUT
+Too many arguments, starting with 'unknown1'
+
+
+  --help           Show full help page.
+
+  <argument-one>   First argument
+  (required)
+
+  <argument-two>   Second argument
+  (required)
+
+STDERR_OUTPUT,
             ],
             'non-configured-extra-argument-after-double-dash' => [
-                'script'           => __DIR__ . '/scripts/two-arguments.php',
-                'parametersString' => 'asd -- fgh unknown1 unknown2',
-                'errorOutput'      => "Too many arguments, starting with 'unknown1'",
-                'helpSubstrings'   => [
-                    '<argument-one>   First argument',
-                    '<argument-two>   Second argument'
-                ],
+                'scriptPath'          => __DIR__ . '/scripts/two-arguments.php',
+                'parametersString'    => 'asd -- fgh unknown1 unknown2',
+                'expectedErrorOutput' => <<<STDERR_OUTPUT
+Too many arguments, starting with 'unknown1'
+
+
+  --help           Show full help page.
+
+  <argument-one>   First argument
+  (required)
+
+  <argument-two>   Second argument
+  (required)
+
+STDERR_OUTPUT,
             ],
             'non-configured-option-name' => [
-                'script'           => __DIR__ . '/scripts/two-arguments.php',
-                'parametersString' => 'asd fgh --unknown=value',
-                'errorOutput'      => "Unknown option '--unknown'",
-                'helpSubstrings'   => [],
+                'scriptPath'          => __DIR__ . '/scripts/two-arguments.php',
+                'parametersString'    => 'asd fgh --unknown=value',
+                'expectedErrorOutput' => <<<STDERR_OUTPUT
+Unknown option '--unknown'
+
+
+  --help   Show full help page.
+
+STDERR_OUTPUT,
             ],
             'non-configured-option-short-name' => [
-                'script'           => __DIR__ . '/scripts/two-arguments.php',
-                'parametersString' => 'asd fgh -u value',
-                'errorOutput'      => "Unknown option '-u'",
-                'helpSubstrings'   => [],
+                'scriptPath'          => __DIR__ . '/scripts/two-arguments.php',
+                'parametersString'    => 'asd fgh -u value',
+                'expectedErrorOutput' => <<<STDERR_OUTPUT
+Unknown option '-u'
+
+
+  --help   Show full help page.
+
+STDERR_OUTPUT,
             ],
         ];
     }